add trace generation
