+++ prover/src/tests/cpu_tests.rs
@@ -7,7 +7,7 @@
 
 use crate::tables::cpu::{CpuOperation, bus_interactions, cols, generate_cpu_trace};
 use crate::tables::trace_builder::Traces;
-use crate::tables::types::FE;
+use crate::tables::types::{DecodeEntry, FE};
 
 use executor::{
     elf::Elf,
@@ -17,11 +17,12 @@ use executor::{
 /// Helper to create 4 operations from a template (required for power-of-2 trace).
 fn ops4(op: CpuOperation) -> Vec<CpuOperation> {
     (0..4)
-        .map(|i| CpuOperation {
-            timestamp: (i as u64) * 4,
-            pc: op.pc + (i as u64) * 4,
-            next_pc: op.pc + (i as u64) * 4 + 4,
-            ..op.clone()
+        .map(|i| {
+            let mut new_op = op.clone();
+            new_op.timestamp = (i as u64) * 4;
+            new_op.decode.pc = op.decode.pc + (i as u64) * 4;
+            new_op.next_pc = op.decode.pc + (i as u64) * 4 + 4;
+            new_op
         })
         .collect()
 }
@@ -30,8 +31,8 @@ fn ops4(op: CpuOperation) -> Vec<CpuOperation> {
 fn test_cpu_operation_default() {
     let op = CpuOperation::new();
     assert_eq!(op.timestamp, 0);
-    assert_eq!(op.pc, 0);
-    assert!(!op.op_add);
+    assert_eq!(op.decode.pc, 0);
+    assert!(!op.decode.op_add);
     assert!(!op.branch_cond);
 }
 
@@ -39,7 +40,7 @@ fn test_cpu_operation_default() {
 fn test_cpu_operation_compute_arg1_no_extension() {
     let mut op = CpuOperation::new();
     op.rv1 = 0x1234_5678_9ABC_DEF0;
-    op.word_instr = false;
+    op.decode.word_instr = false;
 
     assert_eq!(op.compute_arg1(), 0x1234_5678_9ABC_DEF0);
 }
@@ -48,8 +49,8 @@ fn test_cpu_operation_compute_arg1_no_extension() {
 fn test_cpu_operation_compute_arg1_word_zero_extend() {
     let mut op = CpuOperation::new();
     op.rv1 = 0x1234_5678_9ABC_DEF0;
-    op.word_instr = true;
-    op.signed = false;
+    op.decode.word_instr = true;
+    op.decode.signed = false;
 
     // Should zero-extend from lower 32 bits
     assert_eq!(op.compute_arg1(), 0x9ABC_DEF0);
@@ -59,8 +60,8 @@ fn test_cpu_operation_compute_arg1_word_zero_extend() {
 fn test_cpu_operation_compute_arg1_word_sign_extend_positive() {
     let mut op = CpuOperation::new();
     op.rv1 = 0x1234_5678_1ABC_DEF0; // Positive 32-bit value
-    op.word_instr = true;
-    op.signed = true;
+    op.decode.word_instr = true;
+    op.decode.signed = true;
 
     // Bit 31 is 0, so sign extension keeps it positive
     assert_eq!(op.compute_arg1(), 0x1ABC_DEF0);
@@ -70,8 +71,8 @@ fn test_cpu_operation_compute_arg1_word_sign_extend_positive() {
 fn test_cpu_operation_compute_arg1_word_sign_extend_negative() {
     let mut op = CpuOperation::new();
     op.rv1 = 0x1234_5678_8000_0001; // Negative when viewed as 32-bit signed
-    op.word_instr = true;
-    op.signed = true;
+    op.decode.word_instr = true;
+    op.decode.signed = true;
 
     // Per spec constraint: arg1[4:] = (2^32-1) * rv1_sign_bit * signed
     // For signed word instructions with sign bit set, arg1 is sign-extended.
@@ -82,8 +83,8 @@ fn test_cpu_operation_compute_arg1_word_sign_extend_negative() {
 fn test_cpu_operation_compute_arg2_store() {
     let mut op = CpuOperation::new();
     op.rv2 = 0xDEAD_BEEF;
-    op.imm = 0x1234;
-    op.op_store = true;
+    op.decode.imm = 0x1234;
+    op.decode.op_store = true;
 
     // STORE uses imm for address calculation (addr = rv1 + imm)
     // rv2 is the value being stored, not part of address
@@ -94,8 +95,8 @@ fn test_cpu_operation_compute_arg2_store() {
 fn test_cpu_operation_compute_arg2_load() {
     let mut op = CpuOperation::new();
     op.rv2 = 0xDEAD_BEEF;
-    op.imm = 0x1234;
-    op.op_load = true;
+    op.decode.imm = 0x1234;
+    op.decode.op_load = true;
 
     // LOAD uses imm for address calculation (addr = rv1 + imm)
     assert_eq!(op.compute_arg2(), 0x1234);
@@ -105,8 +106,8 @@ fn test_cpu_operation_compute_arg2_load() {
 fn test_cpu_operation_compute_arg2_beq() {
     let mut op = CpuOperation::new();
     op.rv2 = 0xCAFE_BABE;
-    op.imm = 0x5678;
-    op.op_beq = true;
+    op.decode.imm = 0x5678;
+    op.decode.op_beq = true;
 
     // BEQ uses rv2
     assert_eq!(op.compute_arg2(), 0xCAFE_BABE);
@@ -116,9 +117,9 @@ fn test_cpu_operation_compute_arg2_beq() {
 fn test_cpu_operation_compute_arg2_add_with_imm() {
     let mut op = CpuOperation::new();
     op.rv2 = 0;
-    op.rs2 = 0; // rs2 = 0 means use immediate
-    op.imm = 0x1234_5678;
-    op.op_add = true;
+    op.decode.rs2 = 0; // rs2 = 0 means use immediate
+    op.decode.imm = 0x1234_5678;
+    op.decode.op_add = true;
 
     // ADD with rs2=0 uses imm
     assert_eq!(op.compute_arg2(), 0x1234_5678);
@@ -128,9 +129,9 @@ fn test_cpu_operation_compute_arg2_add_with_imm() {
 fn test_cpu_operation_compute_arg2_add_with_rs2() {
     let mut op = CpuOperation::new();
     op.rv2 = 0xABCD_EF00;
-    op.rs2 = 5; // Non-zero rs2
-    op.imm = 0x1234_5678;
-    op.op_add = true;
+    op.decode.rs2 = 5; // Non-zero rs2
+    op.decode.imm = 0x1234_5678;
+    op.decode.op_add = true;
 
     // ADD with rs2 != 0 uses rv2
     assert_eq!(op.compute_arg2(), 0xABCD_EF00);
@@ -153,12 +154,15 @@ fn test_sign_bit_32_negative() {
 #[test]
 fn test_trace_generation_basic() {
     let ops = ops4(CpuOperation {
-        pc: 0x1000,
-        rs1: 1,
-        rs2: 2,
-        rd: 3,
-        write_register: true,
-        op_add: true,
+        decode: DecodeEntry {
+            pc: 0x1000,
+            rs1: 1,
+            rs2: 2,
+            rd: 3,
+            write_register: true,
+            op_add: true,
+            ..Default::default()
+        },
         rv1: 10,
         rv2: 20,
         res: 30,
@@ -187,8 +191,11 @@ fn test_trace_generation_basic() {
 #[test]
 fn test_trace_generation_64bit_pc() {
     let ops = ops4(CpuOperation {
-        pc: 0x8000_0000_1234_5678,
-        op_add: true,
+        decode: DecodeEntry {
+            pc: 0x8000_0000_1234_5678,
+            op_add: true,
+            ..Default::default()
+        },
         ..Default::default()
     });
 
@@ -206,8 +213,11 @@ fn test_trace_generation_64bit_pc() {
 #[test]
 fn test_trace_generation_rv1_dwordwhh() {
     let ops = ops4(CpuOperation {
+        decode: DecodeEntry {
+            op_add: true,
+            ..Default::default()
+        },
         rv1: 0xFFFF_EEEE_DDDD_CCCCu64,
-        op_add: true,
         ..Default::default()
     });
 
@@ -223,9 +233,12 @@ fn test_trace_generation_rv1_dwordwhh() {
 #[test]
 fn test_trace_generation_arg1_dwordbl() {
     let ops = ops4(CpuOperation {
+        decode: DecodeEntry {
+            word_instr: false,
+            op_add: true,
+            ..Default::default()
+        },
         rv1: 0x0807_0605_0403_0201u64,
-        word_instr: false,
-        op_add: true,
         ..Default::default()
     });
 
@@ -249,8 +262,11 @@ fn test_trace_generation_res_dwordbl() {
     // Set rv1 to the desired result value since arg1 = rv1 when word_instr=false,
     // and arg2 = 0 (imm default) when rs2=0.
     let ops = ops4(CpuOperation {
+        decode: DecodeEntry {
+            op_add: true,
+            ..Default::default()
+        },
         rv1: 0xFEDC_BA98_7654_3210u64,
-        op_add: true,
         ..Default::default()
     });
 
@@ -272,10 +288,13 @@ fn test_trace_generation_res_dwordbl() {
 #[test]
 fn test_trace_generation_sign_bits() {
     let ops = ops4(CpuOperation {
+        decode: DecodeEntry {
+            word_instr: true,
+            op_add: true,
+            ..Default::default()
+        },
         rv1: 0x0000_0000_8000_0000u64, // bit 31 set
         res: 0x0000_0000_8000_0000u64, // bit 31 set
-        word_instr: true,
-        op_add: true,
         ..Default::default()
     });
 
@@ -384,15 +403,15 @@ fn test_cpu_operation_from_log_arith() {
         dst_val: 300,
     };
 
-    let op = CpuOperation::from_log(&log, 0, instruction);
+    let op = CpuOperation::from_log_and_instruction(&log, 0, instruction);
 
-    assert_eq!(op.pc, 0x1000);
+    assert_eq!(op.decode.pc, 0x1000);
     assert_eq!(op.next_pc, 0x1004);
-    assert_eq!(op.rd, 10);
-    assert_eq!(op.rs1, 11);
-    assert_eq!(op.rs2, 12);
-    assert!(op.op_add);
-    assert!(op.write_register);
+    assert_eq!(op.decode.rd, 10);
+    assert_eq!(op.decode.rs1, 11);
+    assert_eq!(op.decode.rs2, 12);
+    assert!(op.decode.op_add);
+    assert!(op.decode.write_register);
     assert_eq!(op.rv1, 100);
     assert_eq!(op.rv2, 200);
     assert_eq!(op.res, 300);
@@ -418,12 +437,12 @@ fn test_cpu_operation_from_log_branch() {
         dst_val: 0,
     };
 
-    let op = CpuOperation::from_log(&log, 4, instruction);
+    let op = CpuOperation::from_log_and_instruction(&log, 4, instruction);
 
     assert_eq!(op.timestamp, 4);
-    assert_eq!(op.pc, 0x2000);
-    assert!(op.op_blt);
-    assert!(op.signed);
+    assert_eq!(op.decode.pc, 0x2000);
+    assert!(op.decode.op_blt);
+    assert!(op.decode.signed);
     assert!(op.branch_cond); // 10 < 20
     // For BLT, res is the comparison result (0 or 1), not subtraction
     // res[0] = 1 if arg1 < arg2, res[1..7] = 0 (enforced by SLT res zero constraint)
@@ -450,8 +469,8 @@ fn test_cpu_operation_from_log_word_instr() {
         dst_val: 0xFFFF_FFFF_8000_0001, // Result sign-extended
     };
 
-    let op = CpuOperation::from_log(&log, 8, instruction);
+    let op = CpuOperation::from_log_and_instruction(&log, 8, instruction);
 
-    assert!(op.word_instr);
-    assert!(op.op_add);
+    assert!(op.decode.word_instr);
+    assert!(op.decode.op_add);
 }
